+++ src/ops/mod.rs
@@ -1,3 +1,4 @@
+pub mod actor_merge;
 pub mod check;
 pub mod clean;
 pub mod import;